+++ Test/StandardUtils.Unit.Tests/Helpers/CryptoHelperTests.cs
@@ -1,10 +1,9 @@
 using System;
 
 using NUnit.Framework;
-
 using StandardUtils.Helpers;
 
-namespace StandardUtils.Unit.Tests
+namespace StandardUtils.Unit.Tests.Helpers
 {
    [TestFixture]
     public class CryptoHelperTests
