+++ Test/StandardUtils.Unit.Tests/Helpers/GuidHelperTests.cs
@@ -1,10 +1,8 @@
 using System;
-
 using NUnit.Framework;
-
 using StandardUtils.Helpers;
 
-namespace StandardUtils.Unit.Tests
+namespace StandardUtils.Unit.Tests.Helpers
 {
     [TestFixture]
     public class GuidHelperTests
